Add Day 10 Part 02 Device
Pretty confident this works, but I can't actually use it to computer the full answer because the script runs too long on the device, even with sleeps.

I could precompute the path and just run the raycasting portion, but I also can only output 100 log lines per run and need to do like 12,000, so...tabling this for now. Will return when I want this star.